+++ src/ESPRotary.cpp
@@ -14,14 +14,21 @@ int ESPRotary::_nextID = 0;
 
 /////////////////////////////////////////////////////////////////
 
+void ESPRotary::_setID() {
+  id = _nextID;
+  _nextID++;
+}
+
+/////////////////////////////////////////////////////////////////
+
 ESPRotary::ESPRotary() {
-  id = _nextID++;
+  _setID();
 }
 
 /////////////////////////////////////////////////////////////////
 
 ESPRotary::ESPRotary(byte pin1, byte pin2, byte steps_per_click /* = 1 */,  int lower_bound /* = INT16_MIN */, int upper_bound /* = INT16_MAX */, int inital_pos /* = 0 */, int increment /* = 1 */) {
-  id = _nextID++;
+  ESPRotary();
   begin(pin1, pin2, steps_per_click, lower_bound, upper_bound, inital_pos, increment);
 }