+++ src/ESPRotary.h
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////
 /*
   ESP8266/Arduino Library for reading rotary encoder values.
-  Copyright 2017-2021 Lennart Hennigs
+  Copyright 2017-2022 Lennart Hennigs
 */
 /////////////////////////////////////////////////////////////////
 #pragma once
@@ -15,10 +15,19 @@
 
 /////////////////////////////////////////////////////////////////
 
-enum direction {
+enum rotary_direction {
   undefined = 0,
   right = 1,
   left = 255
+};
+
+enum rotary_event {
+  left_rotation,
+  right_rotation,
+  speedup_triggered,
+  upper_bound_hit,
+  lower_bound_hit,
+  none
 };
 
 /////////////////////////////////////////////////////////////////
@@ -28,11 +37,20 @@ class ESPRotary {
     int id;
     byte pin1, pin2;
     byte steps_per_click;
-    int lower_bound, upper_bound;
-    int position, last_position;
-    direction dir;
+    int lower_bound;
+    int upper_bound;
     byte state;
     int increment;
+    int steps = 0;  
+    int last_steps = 0;
+    rotary_event last_event;
+    rotary_direction dir;
+    
+    bool retrigger_event = true;
+    bool enable_speedup = false;
+    int speedup_increment = 5;
+    int speedup_interval = 75;
+    long last_turn = 0;
 
     typedef void (*CallbackFunction) (ESPRotary&);
     CallbackFunction change_cb = NULL;
@@ -40,6 +58,7 @@ class ESPRotary {
     CallbackFunction left_cb = NULL;
     CallbackFunction lower_cb = NULL;
     CallbackFunction upper_cb = NULL;
+    CallbackFunction speedup_cb = NULL;
 
   public:
     ESPRotary();
@@ -50,11 +69,21 @@ class ESPRotary {
     int getPosition() const;
     void resetPosition(int p = 0, bool fireCallback = true);
 
-    direction getDirection() const;
-    String directionToString(direction dir) const;
+    rotary_direction getDirection() const;
+    String directionToString(rotary_direction dir) const;
 
     void setIncrement(int inc);
     int getIncrement() const;
+
+    void enableSpeedup(bool enable);
+    void setSpeedupInterval(int time);
+    void setSpeedupIncrement(int inc);
+    int getSpeedupInterval() const;
+    int getSpeedupIncrement() const;
+    bool isSpeedupEnabled() const;
+
+    rotary_event getLastEvent() const;
+    void retriggerEvent(bool retrigger);
 
     void setUpperBound(int upper_bound);
     void setLowerBound(int lower_bound);
@@ -69,6 +98,7 @@ class ESPRotary {
     void setLeftRotationHandler(CallbackFunction f);
     void setUpperOverflowHandler(CallbackFunction f);
     void setLowerOverflowHandler(CallbackFunction f);
+    void setSpeedupHandler(CallbackFunction f);
 
     int getID() const;
     void setID(int newID);
@@ -79,7 +109,11 @@ class ESPRotary {
 
   private:
     static int _nextID;
-
+    
+    void _setEvent(rotary_event e);
+    bool _wasRotated();
+    bool _isWithinBounds(bool alert = false);
+    void _checkForSpeedup(long now);
     void _setID();
 };