+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added ESP8266 and ESP32 interrupt examples
+- fixed bug with ID
 
 ## [1.6.1] - 2022-12-24